third_party/blink/web_tests/fast/forms/datalist/datalist-child-validation-expected.txt

Test for child elements of a datalist element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS e.willValidate is false
PASS e.willValidate is true
PASS w.querySelector(":invalid") is e
PASS e.willValidate is true
PASS e.willValidate is true
PASS document.querySelector(":invalid") is e
PASS document.getElementById("inLegend").willValidate is false
PASS successfullyParsed is true

TEST COMPLETE
